Added base stump of branchless unpacking

+++ src/gpu-common/gpu-device-types.cuh
@@ -3,7 +3,7 @@
 #ifndef GPU_DEVICE_TYPES_CUH
 #define GPU_DEVICE_TYPES_CUH
 
-using vbw_t = uint8_t;
+using vbw_t = uint32_t;
 using lane_t = uint16_t;
 using si_t = uint16_t; // si = Start Index
 using vi_t = uint32_t; // vi = Vector Index
